+++ dbeam-core/src/main/scala/com/spotify/dbeam/args/JdbcConnectionArgs.java
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.spotify.dbeam.options;
+package com.spotify.dbeam.args;
 
 import com.google.auto.value.AutoValue;
 import com.google.common.base.Preconditions;
@@ -29,7 +29,7 @@ import javax.annotation.Nullable;
  * A POJO describing a how to create a JDBC {@link Connection}.
  */
 @AutoValue
-public abstract class JdbcConnectionConfiguration implements Serializable {
+public abstract class JdbcConnectionArgs implements Serializable {
   public abstract String driverClassName();
   public abstract String url();
   @Nullable abstract String username();
@@ -43,26 +43,26 @@ public abstract class JdbcConnectionConfiguration implements Serializable {
     abstract Builder setUrl(String url);
     abstract Builder setUsername(String username);
     abstract Builder setPassword(String password);
-    abstract JdbcConnectionConfiguration build();
+    abstract JdbcConnectionArgs build();
   }
 
-  public static JdbcConnectionConfiguration create(String url)
+  public static JdbcConnectionArgs create(String url)
       throws ClassNotFoundException {
     Preconditions.checkArgument(url != null,
                   "DataSourceConfiguration.create(driverClassName, url) called "
                   + "with null url");
     final String driverClassName = JdbcConnectionUtil.getDriverClass(url);
-    return new AutoValue_JdbcConnectionConfiguration.Builder()
+    return new AutoValue_JdbcConnectionArgs.Builder()
         .setDriverClassName(driverClassName)
         .setUrl(url)
         .build();
   }
 
-  public JdbcConnectionConfiguration withUsername(String username) {
+  public JdbcConnectionArgs withUsername(String username) {
     return builder().setUsername(username).build();
   }
 
-  public JdbcConnectionConfiguration withPassword(String password) {
+  public JdbcConnectionArgs withPassword(String password) {
     return builder().setPassword(password).build();
   }
