+++ dbeam-core/src/main/scala/com/spotify/dbeam/avro/JdbcAvroRecord.java
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.spotify.dbeam;
+package com.spotify.dbeam.avro;
 
 import com.google.common.collect.ImmutableMap;
