+++ dbeam-core/src/main/scala/com/spotify/dbeam/options/JdbcExportArgsFactory.java
@@ -18,7 +18,17 @@ package com.spotify.dbeam.options;
 
 import com.google.common.base.Preconditions;
 
+import com.spotify.dbeam.args.JdbcAvroArgs;
+import com.spotify.dbeam.args.JdbcConnectionArgs;
+import com.spotify.dbeam.args.JdbcExportArgs;
+import com.spotify.dbeam.args.QueryBuilderArgs;
+
 import org.apache.beam.sdk.options.PipelineOptions;
+import org.joda.time.DateTime;
+import org.joda.time.Days;
+import org.joda.time.Period;
+import org.joda.time.ReadablePeriod;
+import org.joda.time.format.ISODateTimeFormat;
 
 import java.io.IOException;
 import java.util.Optional;
@@ -32,7 +42,7 @@ public class JdbcExportArgsFactory {
                                 "'connectionUrl' must be defined");
 
     final JdbcAvroArgs jdbcAvroArgs = JdbcAvroArgs.create(
-        JdbcConnectionConfiguration.create(exportOptions.getConnectionUrl())
+        JdbcConnectionArgs.create(exportOptions.getConnectionUrl())
             .withUsername(exportOptions.getUsername())
             .withPassword(PasswordReader.readPassword(exportOptions).orElse(null)),
         exportOptions.getFetchSize(),
@@ -40,11 +50,51 @@ public class JdbcExportArgsFactory {
 
     return JdbcExportArgs.create(
         jdbcAvroArgs,
-        QueryBuilderArgs.create(exportOptions),
+        createQueryArgs(exportOptions),
         exportOptions.getAvroSchemaNamespace(),
         Optional.ofNullable(exportOptions.getAvroDoc()),
         exportOptions.isUseAvroLogicalTypes()
     );
+  }
+
+  public static QueryBuilderArgs createQueryArgs(JdbcExportPipelineOptions options) {
+    final ReadablePeriod partitionPeriod = Optional.ofNullable(options.getPartitionPeriod())
+        .map(v -> (ReadablePeriod) Period.parse(v)).orElse(Days.ONE);
+    Optional<DateTime> partition = Optional.ofNullable(options.getPartition())
+        .map(JdbcExportArgsFactory::parseDateTime);
+    Optional<String> partitionColumn = Optional.ofNullable(options.getPartitionColumn());
+    Preconditions.checkArgument(!partitionColumn.isPresent() || partition.isPresent(),
+                                "To use --partitionColumn the --partition parameter must also be configured");
+
+    if (!(options.isSkipPartitionCheck() || partitionColumn.isPresent())) {
+      DateTime minPartitionDateTime = Optional.ofNullable(options.getMinPartitionPeriod())
+          .map(JdbcExportArgsFactory::parseDateTime)
+          .orElse(DateTime.now().minus(partitionPeriod.toPeriod().multipliedBy(2)));
+      partition.map(p -> validatePartition(p, minPartitionDateTime));
+    }
+    return QueryBuilderArgs.create(options.getTable())
+        .builder()
+        .setLimit(Optional.ofNullable(options.getLimit()))
+        .setPartitionColumn(partitionColumn)
+        .setPartition(partition)
+        .setPartitionPeriod(partitionPeriod)
+        .build();
+  }
+
+  private static DateTime parseDateTime(String input) {
+    if (input.endsWith("Z")) {
+      input = input.substring(0, input.length() -1);
+    }
+    return DateTime.parse(input, ISODateTimeFormat.localDateOptionalTimeParser());
+  }
+
+  private static DateTime validatePartition(DateTime partitionDateTime, DateTime minPartitionDateTime) {
+    Preconditions.checkArgument(
+        partitionDateTime.isAfter(minPartitionDateTime),
+        "Too old partition date %s. Use a partition date >= %s or use --skip-partition-check",
+        partitionDateTime, minPartitionDateTime
+    );
+    return partitionDateTime;
   }
 
 }
